requirements.txt
pandas>=2.0.0
openai>=1.0.0
customtkinter>=5.2.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
